Factor out key translation into plugin, use events instead to fix key repeat bug
